# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Social Connect API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: number;
  clerkId: string;
  username: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  coverUrl?: string | null;
  friendCount?: number;
  groupCount?: number;
  createdAt: string;
}

export interface UserProfileUpdate {
  /** @minLength 3 */
  username?: string;
  /** @minLength 1 */
  displayName?: string;
  bio?: string;
  /**
     * @minimum 13
     * @maximum 120
     */
  age?: number;
  avatarUrl?: string;
  coverUrl?: string;
}

/**
 * @nullable
 */
export type UserSummaryFriendStatus = typeof UserSummaryFriendStatus[keyof typeof UserSummaryFriendStatus] | null;


export const UserSummaryFriendStatus = {
  friends: 'friends',
  pending_sent: 'pending_sent',
  pending_received: 'pending_received',
} as const;

export interface UserSummary {
  id: number;
  clerkId: string;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  friendStatus?: UserSummaryFriendStatus;
}

export type FriendRequestStatus = typeof FriendRequestStatus[keyof typeof FriendRequestStatus];


export const FriendRequestStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
} as const;

export interface FriendRequest {
  id: number;
  fromUserId: number;
  toUserId: number;
  status: FriendRequestStatus;
  fromUser?: UserSummary;
  toUser?: UserSummary;
  createdAt: string;
}

/**
 * @nullable
 */
export type MessageMediaType = typeof MessageMediaType[keyof typeof MessageMediaType] | null;


export const MessageMediaType = {
  image: 'image',
  video: 'video',
} as const;

export interface Message {
  id: number;
  senderId: number;
  /** @nullable */
  recipientId?: number | null;
  /** @nullable */
  groupId?: number | null;
  content: string;
  /** @nullable */
  mediaUrl?: string | null;
  /** @nullable */
  mediaType?: MessageMediaType;
  sender?: UserSummary;
  createdAt: string;
}

export type MessageInputMediaType = typeof MessageInputMediaType[keyof typeof MessageInputMediaType];


export const MessageInputMediaType = {
  image: 'image',
  video: 'video',
} as const;

export interface MessageInput {
  /** @minLength 1 */
  content: string;
  mediaUrl?: string;
  mediaType?: MessageInputMediaType;
}

export interface ConversationSummary {
  user: UserSummary;
  lastMessage: Message;
  unreadCount: number;
}

/**
 * @nullable
 */
export type GroupMyRole = typeof GroupMyRole[keyof typeof GroupMyRole] | null;


export const GroupMyRole = {
  admin: 'admin',
  member: 'member',
} as const;

export interface Group {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  isPublic: boolean;
  memberCount: number;
  /** @nullable */
  myRole?: GroupMyRole;
  createdAt: string;
}

export interface GroupInput {
  /** @minLength 2 */
  name: string;
  description?: string;
  avatarUrl?: string;
  isPublic?: boolean;
}

export interface GroupUpdate {
  /** @minLength 2 */
  name?: string;
  description?: string;
  avatarUrl?: string;
  isPublic?: boolean;
}

export type GroupMemberRole = typeof GroupMemberRole[keyof typeof GroupMemberRole];


export const GroupMemberRole = {
  admin: 'admin',
  member: 'member',
} as const;

export interface GroupMember {
  user: UserSummary;
  role: GroupMemberRole;
  joinedAt: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}

export type SearchUsersParams = {
q: string;
limit?: number;
};

export type GetDirectMessagesParams = {
before?: number;
limit?: number;
};

export type SearchGroupsParams = {
q: string;
limit?: number;
};

export type GetGroupMessagesParams = {
before?: number;
limit?: number;
};