# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Social Connect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /friends/requests/{userId}:
    post:
      operationId: sendFriendRequest
      tags: [friends]
      summary: Send a friend request
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Request sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendRequest"
    delete:
      operationId: cancelFriendRequest
      tags: [friends]
      summary: Cancel or unfriend
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
  /friends/requests/{requestId}/accept:
    post:
      operationId: acceptFriendRequest
      tags: [friends]
      summary: Accept a friend request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendRequest"
  /friends/requests/{requestId}/reject:
    post:
      operationId: rejectFriendRequest
      tags: [friends]
      summary: Reject a friend request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendRequest"

  # ── Direct Messages ───────────────────────────────────────────────────────────
...
  /groups:
    get:
      operationId: getMyGroups
      tags: [groups]
      summary: Get groups the current user belongs to
      responses:
        "200":
          description: Groups list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a new group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/search:
    get:
      operationId: searchGroups
      tags: [groups]
      summary: Search for public groups
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Matching groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
  /groups/{groupId}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get group details
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "404":
          description: Not found
    patch:
      operationId: updateGroup
      tags: [groups]
      summary: Update group info (admin only)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupUpdate"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
    delete:
      operationId: deleteGroup
      tags: [groups]
      summary: Delete a group (admin only)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /groups/{groupId}/join:
    post:
      operationId: joinGroup
      tags: [groups]
      summary: Join a public group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{groupId}/leave:
    post:
      operationId: leaveGroup
      tags: [groups]
      summary: Leave a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Left
  /groups/{groupId}/members:
    get:
      operationId: getGroupMembers
      tags: [groups]
      summary: Get members of a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Member list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupMember"
  /groups/{groupId}/messages:
    get:
      operationId: getGroupMessages
      tags: [messages]
      summary: Get messages in a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: before
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendGroupMessage
      tags: [messages]
      summary: Send a message to a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ── Storage ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    UserSummary:
      type: object
      required: [id, clerkId, username, displayName]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        friendStatus:
          type: ["string", "null"]
          enum: ["friends", "pending_sent", "pending_received", null]
    FriendRequest:
      type: object
      required: [id, fromUserId, toUserId, status, createdAt]
      properties:
        id:
          type: integer
        fromUserId:
          type: integer
        toUserId:
          type: integer
        status:
          type: string
          enum: [pending, accepted, rejected]
        fromUser:
          $ref: "#/components/schemas/UserSummary"
        toUser:
          $ref: "#/components/schemas/UserSummary"
        createdAt:
          type: string
          format: date-time
    Message:
      type: object
      required: [id, senderId, content, createdAt]
      properties:
        id:
          type: integer
        senderId:
          type: integer
        recipientId:
          type: ["integer", "null"]
        groupId:
          type: ["integer", "null"]
        content:
          type: string
        mediaUrl:
          type: ["string", "null"]
        mediaType:
          type: ["string", "null"]
          enum: ["image", "video", null]
        sender:
          $ref: "#/components/schemas/UserSummary"
        createdAt:
          type: string
          format: date-time
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        mediaUrl:
          type: string
        mediaType:
          type: string
          enum: [image, video]
...
    Group:
      type: object
      required: [id, name, isPublic, memberCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isPublic:
          type: boolean
        memberCount:
          type: integer
        myRole:
          type: ["string", "null"]
          enum: ["admin", "member", null]
        createdAt:
          type: string
          format: date-time
    GroupInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
        description:
          type: string
        avatarUrl:
          type: string
        isPublic:
          type: boolean
          default: true
    GroupUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        description:
          type: string
        avatarUrl:
          type: string
        isPublic:
          type: boolean
    GroupMember:
      type: object
      required: [user, role, joinedAt]
      properties:
        user:
          $ref: "#/components/schemas/UserSummary"
        role:
          type: string
          enum: [admin, member]
        joinedAt:
          type: string
          format: date-time
...